add task queue test cases
