+++ test/AElf.Core.Tests/TaskQueueTests.cs
@@ -1,123 +1,72 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using NSubstitute.ExceptionExtensions;
 using Shouldly;
 using Xunit;
 
 namespace AElf
 {
-    public class TaskQueueTests : CoreAElfTestBase
+    public class TaskQueueTests
     {
-        private readonly ITaskQueueManager _taskQueueManager;
-
+        private TaskQueue _taskQueue;
+        private int _counter;
         public TaskQueueTests()
         {
-            _taskQueueManager = GetRequiredService<ITaskQueueManager>();
+            _taskQueue = new TaskQueue();
+            _counter = 0;
         }
 
+        [Fact]
+        public async Task StartQueueTest_Twice()
+        {
+            _taskQueue.Start();
+            Should.Throw<InvalidOperationException>(() => _taskQueue.Start());
+        }
 
         [Fact]
-        public async Task Test_Enqueue()
+        public async Task EnqueueTest()
         {
-            var result = 1;
-            var testQueue = _taskQueueManager.CreateQueue("TestQueue");
-            Parallel.For(0, 100, i =>
+            _taskQueue.Start();
+            _taskQueue.Enqueue(ProcessTask);
+            Thread.Sleep(10);
+            _counter.ShouldBe(1);
+            
+            _taskQueue.Dispose();
+            Should.Throw<InvalidOperationException>(() => _taskQueue.Enqueue(ProcessTask));
+        }
+
+        [Fact]
+        public async Task Enqueue_MultipleTimes()
+        {
+            _taskQueue.Start();
+            for (var i = 0; i < 10; i++)
             {
-                testQueue.Enqueue(async () =>
-                {
-                    var value = result;
-                    result = value + 1;
-                });
-            });
-
-            testQueue.Dispose();
-
-            result.ShouldBe(101);
+                _taskQueue.Enqueue(ProcessTask);
+                Thread.Sleep(10);
+            }
+            _counter.ShouldBe(10);
         }
 
         [Fact]
-        public async Task Test_Many_Enqueue()
+        public async Task Dispose_QueueTest()
         {
-            var testData = new int[3];
-            var testQueueA = _taskQueueManager.CreateQueue("TestQueueA");
-            var testQueueB = _taskQueueManager.CreateQueue("TestQueueB");
-            var testQueueC = _taskQueueManager.CreateQueue("TestQueueC");
+            _taskQueue.Start();
+            _taskQueue.Dispose();
 
-            Parallel.For(0, 100, i =>
-            {
-                testQueueA.Enqueue(async () => { testData[0]++; });
-                testQueueB.Enqueue(async () => { testData[1]++; });
-                testQueueC.Enqueue(async () => { testData[2]++; });
-            });
-
-            testQueueA.Dispose();
-            testQueueB.Dispose();
-            testQueueC.Dispose();
-
-            testData[0].ShouldBe(100);
-            testData[1].ShouldBe(100);
-            testData[2].ShouldBe(100);
+            Should.Throw<InvalidOperationException>(() => _taskQueue.Enqueue(ProcessTask));
         }
 
         [Fact]
-        public async Task Test_Dispose()
+        public async Task Stop_QueueTest()
         {
-            var result = 1;
-
-            var testQueue = _taskQueueManager.CreateQueue("TestQueue");
-
-            Parallel.For(0, 3, i =>
-            {
-                testQueue.Enqueue(async () =>
-                {
-                    var value = result;
-                    await Task.Delay(100);
-                    result = value + 1;
-                });
-            });
-            testQueue.Dispose();
-
-            result.ShouldBe(4);
-
-            Assert.Throws<InvalidOperationException>(() => testQueue.Enqueue(async () => { result++; }));
+            _taskQueue.Start();
+            _taskQueue.Dispose();
+            Should.Throw<InvalidOperationException>(() => _taskQueue.Enqueue(ProcessTask));
         }
 
-        [Fact]
-        public void Test_GetQueue()
+        private async Task ProcessTask()
         {
-            var defaultQueueA = _taskQueueManager.GetQueue();
-            var defaultQueueB = _taskQueueManager.GetQueue();
-            defaultQueueA.ShouldBe(defaultQueueB);
-
-            var testQueueA1 = _taskQueueManager.CreateQueue("TestQueueA");
-            var testQueueA2 = _taskQueueManager.GetQueue("TestQueueA");
-            testQueueA1.ShouldBe(testQueueA2);
-            testQueueA1.ShouldNotBe(defaultQueueA);
-
-            var testQueueB1 = _taskQueueManager.CreateQueue("TestQueueB");
-            var testQueueB2 = _taskQueueManager.GetQueue("TestQueueB");
-            testQueueB1.ShouldBe(testQueueB2);
-            testQueueB1.ShouldNotBe(defaultQueueA);
-            testQueueB1.ShouldNotBe(testQueueA1);
-        }
-
-        [Fact]
-        public async Task Test_TaskQueueManager_Dispose()
-        {
-            _taskQueueManager.Dispose();
-
-            var testQueue = _taskQueueManager.GetQueue();
-
-            var result = 1;
-            Should.Throw<InvalidOperationException>(() =>
-                testQueue.Enqueue(async () =>
-                {
-                    var value = result;
-                    result = value + 1;
-                })
-            );
+            await Task.Run(() => _counter++);
         }
     }
 }
